+++ Defer.ts
@@ -3,22 +3,27 @@
 /** Defer - functions for strongly typed promise/deferred handling
  */
 class Defer {
+    public static promiseImpl: {
+        defer<T, F>(): PsDeferred<T, F>;
+        resolve<T = any, F = never>(value: T): PsPromise<T, F>;
+        reject<T = never, F = any>(error: F): PsPromise<T, F>;
+        all<T, F>(promises: PsPromise<T, F>[]): PsPromise<T[], F>;
+    } = <{ defer: any; resolve: any; reject: any; all: any }>Q;
 
-    /** Create a Q deferred object with a 'promise' property
-     * @return a PS deferred object with success and error return values
-     */
+
+    /** Create a deferred object with a 'promise' property and 'resolve' and 'reject' methods */
     static newDefer<T, F>(): PsDeferred<T, F> {
-        return <PsDeferred<T, F>>Q.defer<any>();
+        return Defer.promiseImpl.defer<T, F>();
     }
 
 
     static resolve<T = any, F = never>(resolveValue: T): PsPromise<T, F> {
-        return <PsPromise<T, F>>Q.resolve<any>(resolveValue);
+        return Defer.promiseImpl.resolve<T, F>(resolveValue);
     }
 
 
     static reject<T = never, F = any>(rejectValue: F): PsPromise<T, F> {
-        return <PsPromise<T, F>>Q.reject<any>(rejectValue);
+        return Defer.promiseImpl.reject<T, F>(rejectValue);
     }
 
 
@@ -41,7 +46,7 @@ class Defer {
     /** Wait for an array of promises to resolve and return the result in a strongly typed array */
     static when<T, F>(promises: PsPromise<T, F>[]): PsPromise<T[], F>;
     static when<T, F>(promises: PsPromise<T, F>[]): PsPromise<T[], F> {
-        return <PsPromise<T[], F>>Q.all(<any[]>promises);
+        return <PsPromise<T[], F>>Defer.promiseImpl.all(<any[]>promises);
     }
 
 
@@ -51,47 +56,49 @@ class Defer {
     }
 
 
-    /** Takes an argument and throws it, useful for inferring the type of TypeScript promises without having to explicitly give the type */
+    /** Takes an argument and throws it, useful for inferring the type of TypeScript promises without having to explicitly give the type
+     * (Same as throw() but for earlier versions of TypeScript when throw was a keyword not allowed as a method name)
+     */
     static throwBack<F>(error: F): Throws<F> {
         throw error;
     }
 
 
     /** Run each object from 'args' through 'actionFunc' and return a deferred promise that completes when all of the actions complete
-     * @param args: an array of objects to pass individually to 'actionFunc''
-     * @param actionFunc: this action is called with a unique deferred promise that must be resolved or rejected
+     * @param args an array of objects to pass individually to 'actionFunc''
+     * @param actionFunc this action is called with a unique deferred promise that must be resolved or rejected
      * somewhere in the action, and each object from 'args' as a parameter
-     * @param failOnFirstError: true to stop running the actions when the first one throws an error,
+     * @param failOnFirstError true to stop running the actions when the first one throws an error,
      * else continue running and return a list of successful results
      * @return a promise that returns an array of all of the results returned from the calls to 'actionFunc'
      */
     static runActionForAllInSeries<T, R, F>(args: T[], actionFunc: (def: PsDeferred<R, F>, obj: T) => void, stopOnFirstError: boolean = false): PsPromise<R[], F> {
-        var initalDfd = Q.defer<R>();
+        var initalDfd = Defer.promiseImpl.defer<R, F>();
         initalDfd.resolve(<never>null);
         var results: R[] = [];
         var errors: F[] = [];
         // for each action/argument combination, chain it to the previous action result
-        var promise = args.reduce(function runActionForArgInSeries(promise: Q.Promise<R>, arg: T) {
+        var promise = args.reduce(function runActionForArgInSeries(promise: PsPromise<R, F>, arg: T) {
             function successCallNextAction(res: R) {
                 results.push(res);
-                var dfd = <PsDeferred<R, F>>Q.defer<R>();
+                var dfd = Defer.promiseImpl.defer<R, F>();
                 actionFunc(dfd, arg);
-                return <Q.Promise<R> & Q.IWhenable<R>>dfd.promise;
+                return dfd.promise;
             }
 
             function failureCallNextAction(err: F) {
                 errors.push(err);
-                var dfd = <PsDeferred<R, F>>Q.defer<R>();
+                var dfd = Defer.promiseImpl.defer<R, F>();
                 actionFunc(dfd, arg);
-                return <Q.Promise<R> & Q.IWhenable<R>>dfd.promise;
+                return dfd.promise;
             }
 
             // handle errors if all actions need to run
             if (!stopOnFirstError) {
-                return promise.then(successCallNextAction, failureCallNextAction);
+                return <PsPromise<R, F>>promise.then(successCallNextAction, failureCallNextAction);
             }
             else {
-                return promise.then(successCallNextAction);
+                return <PsPromise<R, F>>promise.then(successCallNextAction);
             }
         }, initalDfd.promise);
 
@@ -104,38 +111,25 @@ class Defer {
     }
 
 
-    /** Chain one deferred to another, so resolve and reject callbacks pass from 'srcPromise' to 'dstPromise'
-     * @param srcPromise: the source promise to listen to via Promise.then()
-     * @param dstPromise: the destination to pipe 'srcPromise' Promise.resolve() and Promise.reject() callbacks to
+    /** Chain one deferred to another, so resolve and reject callbacks pass from 'srcPromise' to 'dstDfd'.
+     * With optional success and failure functions that are called before the 'dstDfd' is resolved/rejected.
+     * @param srcPromise the source promise to listen to via Promise.then()
+     * @param dstDfd the destination to pipe 'srcPromise' Promise.resolve() and Promise.reject() callbacks to
+     * @param successCb optional function to call if 'srcPromise' succeeds,
+     * which can optionally modify the result forwarded to 'dstDfd'
+     * @param failureCb optional function to call if 'srcPromise' fails,
+     * which can optionally modify the error forwarded to 'dstDfd'
      */
-    static chainTo<T, F>(srcPromise: PsPromise<T, F>, dstDfd: PsDeferred<T, F>) {
-        srcPromise.then(function chainedPromiseSuccess(res) {
-            dstDfd.resolve(res);
-        }, function chainedPromiseFailure(err) {
-            dstDfd.reject(err);
-        });
-    }
-
-
-    /** Chain one deferred to another, so resolve and reject callbacks pass from 'srcPromise' to 'dstPromise'.
-     * With optional success and failure functions that are called before the 'dstProimse' is resolved/rejected.
-     * @param srcPromise: the source promise to listen to via Promise.then()
-     * @param dstPromise: the destination to pipe 'srcPromise' Promise.resolve() and Promise.reject() callbacks to
-     * @param [successCall]: optional function to call if 'srcPromise' succeeds,
-     * which can optionally modify the result forwarded to 'dstPromise'
-     * @param [failureCall]: optional function to call if 'srcPromise' fails,
-     * which can optionally modify the error forwarded to 'dstPromise'
-     */
-    static chainToWith<T, F1, F2>(srcPromise: PsPromise<T, F1>, dstDfd: PsDeferred<T, F2>, successCall: (obj: T) => T | null | undefined, failureCall: (err: F1) => F2 | null | undefined) {
+    static chainToWith<T, F1, F2>(srcPromise: PsPromise<T, F1>, dstDfd: PsDeferred<T, F2>, successCb?: ((obj: T) => T | null | undefined) | null, failureCb?: ((err: F1) => F2 | null | undefined) | null) {
         if (srcPromise == null || dstDfd == null) {
-            throw new Error("incorrect usage (" + srcPromise + ", " + dstDfd + ", ...), expected (Q.IPromise, Q.Deferred, ...)");
+            throw new Error("incorrect usage (" + srcPromise + ", " + dstDfd + ", ...), expected (PsPromise, PsDeferred, ...)");
         }
 
-        srcPromise.then(function chainedWithActionPromiseSucess(res) {
-            if (successCall) {
+        srcPromise.then(function chainedPromiseSucess(res) {
+            if (successCb) {
                 var newRes: T | null | undefined = null;
                 try {
-                    newRes = successCall(res);
+                    newRes = successCb(res);
                 } catch (successCallErr) {
                     dstDfd.reject(successCallErr);
                 }
@@ -144,12 +138,12 @@ class Defer {
                 }
             }
             dstDfd.resolve(res);
-        }, function chainedWithActionPromiseFailure(err) {
+        }, function chainedPromiseFailure(err) {
             var newErr: F2 = <never>null;
-            if (failureCall) {
+            if (failureCb) {
                 var tmpErr: F2 | null | undefined = null;
                 try {
-                    tmpErr = failureCall(err);
+                    tmpErr = failureCb(err);
                 } catch (failureCallErr) {
                     tmpErr = failureCallErr;
                 }
@@ -166,9 +160,9 @@ class Defer {
 
 
     /** Caches an asynchronous task that returns a deferred so that subsequent calls to
-     * the task resolve or reject with the initial cached result or error
-     * @param work: the function that performs the work and returns a deferred
-     * @return a function with the same signature as 'work' that the returns a cached deferred
+     * the task resolve or reject with the initial cached result or error.
+     * @param work the function that performs the work and returns a deferred
+     * @return a function with the same signature as 'work' that returns a cached deferred
      */
     static cachedDeferredTask<T, F>(work: () => PsDeferred<T, F>): () => PsDeferred<T, F> {
         var cachedDfd = Defer.newDefer<T, F>();
@@ -208,17 +202,17 @@ class Defer {
 
 
     /** Caches an asynchronous task that returns a promise so that subsequent calls to
-     * the task resolve or reject with the initial cached result or error
-     * @param work: the function that performs the work and returns a promise
-     * @return a function with the same signature as 'work' that the returns a cached promise
+     * the task resolve or reject with the initial cached result or error.
+     * @param work the function that performs the work and returns a promise
+     * @return a function with the same signature as 'work' that returns a cached promise
      */
-    static cachedPromiseTask<T extends Q.IPromise<any> | PsPromise<any, any>>(work: () => T): () => T {
+    static cachedPromiseTask<T extends PromiseLike<any> | PsPromise<any, any>>(work: () => T): () => T {
         var cachedPromise: T = <never>undefined;
 
         function cachedPromiseResolver(): T {
             if (cachedPromise === undefined) {
-                var workPromise = work();
-                cachedPromise = workPromise || null;
+                var pWork = work();
+                cachedPromise = pWork || null;
             }
 
             return cachedPromise;